Changed the way RDI is calculated.
The formula remained the same, but the target deficit is now calculated. We take the target deficit set by the user, but also average in all previous months up to this point, so that by the end of the health plan, whether weight loss or gain, the rate of weight loss will have been obtained as closely as possible.